feat: extract shared pre-launch preview module (_preview.py)
Move show_cook_preview body into _render_pre_launch_preview in a new
_preview.py module. Expose show_cook_preview and show_campaign_preview
as thin wrappers for order and fleet entry points respectively.
Update _order.py lazy import source from _prompts to _preview.
Remove show_cook_preview from _prompts.py.
